feat: added the delete api into the views and updated analytics model.
We also changed our models in this. Especially the analytics one
which didnt have the primary key and hence was creating a record for
each task entry. We consolidated the analytics data and run the migrations
with user_id being the primary key. After this I updated how the edit
and add task apis handled the analytics data for futher use.